part 15: pointer, part 1. I spend a lot of time to support long_ptr until I found that type size of long_ptr is missing

+++ tree.c
@@ -5,7 +5,7 @@
 #include "data.h"
 #include "decl.h"
 
-ASTnode *make_ast_node(int op, int type, ASTnode *left, ASTnode *mid, ASTnode *right, int int_value) {
+ASTnode *make_ast_node(int op, int type, ASTnode *left, ASTnode *mid, ASTnode *right, long int_value) {
 
     ASTnode *node = (ASTnode *) malloc(sizeof(ASTnode));
 
@@ -23,10 +23,10 @@ ASTnode *make_ast_node(int op, int type, ASTnode *left, ASTnode *mid, ASTnode *r
     return node;
 }
 
-ASTnode *make_ast_leaf(int op, int type, int int_value) {
+ASTnode *make_ast_leaf(int op, int type, long int_value) {
     return make_ast_node(op, type, NULL, NULL, NULL, int_value);
 }
 
-ASTnode *make_ast_unary(int op, int type, ASTnode *left, int int_value) {
+ASTnode *make_ast_unary(int op, int type, ASTnode *left, long int_value) {
     return make_ast_node(op, type, left, NULL, NULL, int_value);
 }
